Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Mostly a PoC, unsure if we should bother reviewing/landing, but the one interesting thing it kicked out was MSVC's failure to compile PS#2 without some explicit casts.
> 
> I glanced through it and looked at PS1 -> 3. If you want me to take a closer look, I'd recommend splitting this into smaller pieces. Otherwise, I can just rubber stamp.

I'll probably hold off for a while on this, PDFium is fine as-is, but should the larger chrome codebase adopt an API for smart pointers, then we can revisit.  I just wanted to see if we could poke some holes in the proposed API.  The one gotcha is MSVC unable to handle conversions in ?: expressions.

Patch-set: 3
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0