Update patch set 17
Patch Set 17:

(12 comments)

> Patch Set 15:
> 
> In https://pdfium-review.googlesource.com/62310, we hooked up most (not sure if all?) rendering options, like FPDF_GRAYSCALE, to pdfium_test. Can we do that in this CL as well? Then one can easily run pdfium_test and try out this feature.

I have added 2 options to pdfium_test, --forced-color and --fill-to-stroke. Since it is a sample test I have used a combination of colors which are used as the color scheme when the forced color is set. Please let me know if it will be preferable to to set the individual colors via the command line.

Patch-set: 17
Reviewer: Gerrit User 7586 <7586@fca2d64a-0420-3532-b200-12848c95a47b>