Enable test for formcalc eval().
Then fix the bug in formcalc translation where we were appending
an extraneous NUL onto the end of a WideTextBuf. WideTextBuf's
methods give perfectly valid strings or string views based upon
the accumulated length, so this additional NUL was being passed
into JavaScript as part of the string and then failing to parse
with an invalid token error when the NUL was seen.

- Remove several other occurrences of the same pattern.

Change-Id: I4bb68e573c9e41ecb5eca73b009ce8876238997e