Enable test for formcalc eval().
Then fix the bug in formcalc translation where we were appending
an extraneous NUL onto the end of a WideTextBuf. WideTextBuf's
methods give perfectly valid strings or string views based upon
the accumulated lenght, so this additional NUL was being passed
into JavaScript and failing to parse with an invalid token
error when the NUL was seen.

Change-Id: I4bb68e573c9e41ecb5eca73b009ce8876238997e