Add public API to support forced colors mode
The CL extends support for color scheme in PDFium to a forced
color scheme which is applied when calling the new progressive
render API.
The color scheme is currently applied to only text and path.

For text rendering, the color of the text (fill and stroke) is
changed to the color mentioned in the color scheme.

Similarly colors for fill and stroke can be set for path rendering.
In addition to this there is an option to change the fill paths to
strokes, which is required since with fill paths using a single
color, a graphics with multiple adjacent fill paths will not be
decipherable.

All annotations also follow the above rules, except forms (PDF
annotation widgets like text box, combo box etc.) which
are not part of this change, and will be taken up separately.

There also needs to be a special handling for highlight since
the Multiply blend mode leads to the highlights being less visible
in some color schemes, which will be taken up separately.

The above is achieved by the following changes:

1. Adding a struct for the various graphics objects in a pdf for
which a color can be specified and honored. The
corresponding colors are set by the caller. The CL adds support
for only fill and stroke color of paths and text. Subsequent CLs
will add support for more objects.
2. Adding a new API in the progressive rendering public APIs to
pass in the forced color scheme from the caller.
3. Adding a structure in CPDF_RenderOptions to keep a cache of the
color scheme, and a setter method for the same.
4. Adding a new method CPDF_RenderOptions::TranslateOjectColor()
to translate the color of an object based on its type and
render type to the corresponding forced color.
5. Modifying CPDF_RenderStatus::GetFillArgbInternal() and
CPDF_RenderStatus::GetStrokeArgb() to call the above added method for
translating the color for stroke and fill paths.
6. Modifying CPDF_RenderStatus::ProcessPathPattern() to add code to
convert fill paths to stroke.
7. Adding embedder test for testing rendering using the forced color
scheme in public APIs with different content types.
8. Adding command line flags to pdfium_test to test forced color
scheme rendering using a sample color scheme.

Bug: pdfium:1391
Change-Id: I8b3f43eb0290d11b6a1ee2d5a36dfd1f0c0993e8