Enhance encode_pdf_filter.py to extract PNG IDAT
Adds a new "virtual" filter to encode_pdf_filter.py that extracts the
contents of IDAT chunks from a PNG file. The extracted IDAT data can be
used to "embed" PNG images in a PDF file.

For example, a 640x480 24-bit RGB PNG image might be embedded like so:

  {{object 5 0}} <<
    /Type /XObject
    /Subtype /Image
    /Width 640
    /Height 480
    /BitsPerComponent 8
    /ColorSpace /DeviceRGB
    /Filter /FlateDecode
    /DecodeParms <<
      /Columns 640
      /BitsPerComponent 8
      /Colors 3
      /Predictor 10
    >>
    {{streamlen}}
  >>
  stream
  {{include extracted-png.idat}}

  endstream
  endobj

The filter is "virtual" because it's not an actual PDF filter, although
it does contribute a /FlateDecode pass.

Bug: chromium:587182
Change-Id: Id4f3989e2225627d9d46df1aeba10b60eb04ee75