Enable Skia to render text with two different options.
Currently, rendering text with Skia doesn't support the option to
use optimization for LCD display or disabling anti aliasing, because
CFX_SkiaDeviceDriver::DrawDeviceText() doesn't take the rendering
options as one of the parameters.

This CL adds a parameter |options| of struct CFX_RenderOptions to
DrawDeviceText(), so that we can pass text rendering options into
CFX_SkiaDeviceDriver and affect the rendering results. This change
also makes it possible to add more rendering options for Skia in the
future.

This CL also adds the following two embedder tests for test suite
FPDFViewEmbedderTest:
- RenderHelloWorldWithLcdTextFlags (enabling LCD optimization)
- RenderHelloWorldWithNoSmoothText (disabling anti aliasing)

Bug: pdfium:1522
Change-Id: I7e83939cb3bc40eca624c43eb7b2fa25d7c019b2