Check whether the object is a page node before updating |m_PageList|.
bug_1491.in is rendered blank because when handling the "Goto" action,
the destination points to an object which is a /Pages node instead of
/Page node. This sets |m_PageList[0]| with the wrong object which is
missing contents.

This CL adds a check for valid page object before updating |m_PageList|
in CPDF_Document::GetPageIndex(), so that when processing the "GoTo"
action, |m_PageList[i]| will remain unchanged for page |i| if the
matching object number doesn't point to a /Page node. Later, during the
rendering process, CPDF_Document::GetPageDictionary() can rely on
TraversePDFPages() to update |m_PageList| with more reliable object
numbers.

Bug: pdfium:1491
Change-Id: I4e8a81c6504f52dd5e5b8cd9195c2c60d8885f94