Fix image mask area calculations used by FPDF_RenderPage().
Inside FPDF_RenderPage(), when it encounters an image mask, it falls
back to rasterizing the entire PDF to ensure correctness. There is a
disabled implementation to only rasterize the areas of the page that
contains image masks. This code is disabled because it contains a bug
that causes missing content.

In particular, GetMaskBitmap() calls GetMaskDimensionsAndOffsets() to
figure out where the image mask is. The input for the image mask area
is in the user (page) coordinate space, but the output needs to be in
the device coordinate space. GetMaskDimensionsAndOffsets() does not
always do this transformation correctly, so remove it entirely. Instead,
use CPDF_Page::GetDisplayMatrix() to get the correct transformation
matrix and apply it.

Bug: chromium:674771
Change-Id: Ic827a8f82e3e896e5120f70b51be7bfd149c4ca0