Fix enter/space key press events on Combobox
Currently if user presses enter/space key on a combo box control,
popup menu does not open.

This CL handles the OnChar event in the combobox control.
When the Enter key is pressed, OnChar event returns true and shows the
popup menu if it is not already present.

When the Space key is pressed,
- if the form is non-editable and popup menu is not shown
  popup menu is shown
- if the form is editable
  space is entered as text

Bug: pdfium:1513
Change-Id: I5cc81df41ec925c4ceed6ae040665a146567583f