Update patch set 2
Patch Set 2:

I'm glad we are trying to do this in a small CL in isolation, rather than as part of a (much) larger CL.

To be clear, is the goal of this CL to refactor GetTextRenderOptionsHelper(), i.e. no behavior change, or is trying to fix an issue? If it is fixing an issue, what's the issue exactly?

Patch-set: 2
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>