Add a helper function for creating fill rendering options.
Move the process of creating fill rendering options into a helper
function GetFillRenderOptionsHelper(). This helps distinguish the
difference between fill type (can be 0, FXFILL_ALTERNATE or
FXFILL_WINDING) and other fill option flags, and makes it easier to
replace the integer that represents all fill options with a
struct/class in the future.

Change-Id: Ic43e4daf42179db9575f67ef8def6f64acbdd540