Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > (2 comments)
> > 
> > Kind of a shame to give up using partition alloc's "general" partition under the covers?
> 
> I don't think OpenJPEG is hooked up to PartitionAlloc at all, given the lack of "FXMEM" in third_party/libopenjpeg20/. Also, is there a way to use PartitionAlloc and get aligned allocations? I don't see one.

Ah, Ok.  Thought it was calling into FXMEM eventually.  And yes, getting alinged allocations is more efficient if you can rely on OS tricks, but for our case, the usual trick is a code that allocates N+32 bytes and sticks a pointer to the base address at *(ptr-1) where ptr is the aligned return result.

Patch-set: 1