Make fpdf_annot_unittest and fpdf_editimg_unittest be embeddertests.
These tests are all calling FPDF_CreateNewDocument(), which requires
that the entire PDFium environment be initialized, rather than just
the per-test subsystem initialization that we can get away with in
unittests. Although this happens to work at present, we may get into
trouble down the road when more sophisticated intialization is required
for new features.

Change-Id: I94f144c58924f048fd37ddd6a5d3b56ce3bd12da