Update patch set 1
Patch Set 1:

Motivation:  We're going to need to do the same thing for MakeGarbageCollected, but the "magic" for that involves some internal classes which might change.  So we only want one place to have to change for that.  For parallelism, extend the same concept back to retain ptr.

Not 100% automatic, since linebreaks, etc.

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Work-in-progress: false