Update patch set 2
Patch Set 2:

(1 comment)

The differences here are much more obvious to the eye. For rectangles.pdf, I have to zoom in and look very carefully to see the difference. Here, the differences are noticeable even in the thumbnails on crbug.com/pdfium/1500. In particular, Skia draws a brighter white dot at every intersection, instead of a gray dot.

Patch-set: 2