Reset context after running CXFA_TextLayout::LayoutInternal().
LayoutInternal() runs in a loop, but the conditions for when to reset
the context is not reachable. The caller to LayoutInternal() loops over
index values [0, N), but the check for when the index reaches N is in
LayoutInternal(). As a result, the context never resets and holds
potentially dangling pointers.

Fix this by resetting the context after all the LayoutInternal() calls
finish.

Bug: chromium:997412
Change-Id: I17685e1cce951f35c8b70f8ee820ad81a5cbb209