Update patch set 5
Patch Set 5:

(3 comments)

Hi Lei,

I have updated the code, please take a look at it.

One thing that I observed about the FPDFLink_Enumerate is that the out parameter called start_pos actually returns the  index of the link in the page's annotation vector. The contract is not explicit. Hence I think adding this function still holds value because we would need this function if we have a need to convert the FPDF_LINK to FPDF_ANNOT for getting index in any other scenario than FPDFLink_Enumerate

Thanks
Virender

Patch-set: 5