Update patch set 1
Patch Set 1:

Is this going to lead into having a family of APIs for digital signatures? Do we want put them in a separate header in the public/ directory?

Can two_signatures.pdf be smaller and still have all the document features needed to test digital signatures like the current ~100 KB one?

I imagine trying to handwrite two_signatures.pdf is way too difficult. If we are checking in a PDF, can the CL description explain how it's generated?

Patch-set: 1