Update patch set 1
Patch Set 1:

> Is this going to lead into having a family of APIs for digital signatures? Do we want put them in a separate header in the public/ directory?

Yes, I imagine some 5 more functions till signatures can be verified by client code. One to get the Nth signature, then to access a few properties. Done.

> Can two_signatures.pdf be smaller and still have all the document features needed to test digital signatures like the current ~100 KB one?

I had integration testing in mind, in which case a hexdump of a PKCS#7 signature blob is around 50KB. I tested with Acrobat and LibreOffice, both create signatures with around that size.

> I imagine trying to handwrite two_signatures.pdf is way too difficult. If we are checking in a PDF, can the CL description explain how it's generated?

Actually we can handwrite it for unit testing purposes, given that the PKCS#7 blob can be simply a stub. This instantly solves the size concern as well. Done.

For the reference, I created a self-signed CA + two x509 signing certificates using the openssl commandline tool, then signed an empty PDF with LibreOffice, that's how I created the test file in patch set 1. (Could do the same with Acrobat as well, the output would be quite similar.) But based on the above, this is not needed after all.

Also, FPDF_GetSignatureCount() now returns -1 if the document argument is nullptr.

Patch-set: 1