Enable FPDFAnnotEmbedderTest.AddAndSaveUnderlineAnnotation for Skia/SkiaPaths.
For test FPDFAnnotEmbedderTest.AddAndSaveUnderlineAnnotation, Skia and
SkiaPaths rendering results are acceptable but they don't match AGG's
expected results because they use different methods to load glyphs and
handle anti-alising on drawing paths. Therefore this CL adds the
Skia/SkiaPaths expected checksum value for this embeddertest so that it
can be enabled for Skia and SkiaPaths.

Bug: pdfium:1500
Change-Id: Ibd50de8cd846fe9fedde0400cdc79abb1bc09fb2