Enable gn check for //samples:pdfium_test
gn check doesn't understand conditional includes, so `// nogncheck` is
needed.
If `pdf_enable_skia` is false, it will still complain that the headers
files live under `//skia`, which isn't included as in deps.

Also, add a `//:skia` target so that gn can check it as well.
Otherwise, this error won't be caught when running gn check from a
pdfium checkout, but will surface when running gn check from a chromium
checkout.

Change-Id: I2a6e1dd27dccf681d93b01204da15ccbc824b257