Enable FPDFAnnotEmbedderTest.ModifyRectQuadpointsWithAP for SkiaPaths.
When SkiaPaths is enabled, ModifyRectQuadpointsWithAP's test results
have small discrepancies compared to AGG's due to using different glyph
rendering and anti-aliasing methods. However the test results are
acceptable.

When Skia is enabled, the yellow highlights are missing from all bitmap
rendering results in this test.

This CL adds the expected hash values for SkiaPaths so that this test
can be enabled for SkiaPaths, and keeps the test disabled for Skia until
crbug.com/pdfium/1569 is resolved.

Bug: pdfium:1500
Change-Id: I718d0f0c733488865347f92db68cef7afa339b2e