Add FPDFSignatureObj_GetTime() API
This follows the same pattern as FPDFSignatureObj_GetReason(), so the
client has to call this function twice, but allocation of the buffer
happens outside pdfium.

The buffer format is documented to be NUL-terminated ASCII, since the
string is expected to mostly contain just numbers. An alternative would
be to actually parse the string, but the PDF reference does not go into
details: e.g. it does not specify if the timezone part is mandatory or
not. So probably the best is to leave this to the clients.

Change-Id: I68b981eacd352777452ad5fdd397f7220d891196