Remove CFWL_WidgetProperties::m_pFWLThemeProvider.
The IFWL_ThemeProvider, as implemented by CXFA_FWLThemeProvider is
essentially a singleton living in the CXFA_FFApp. Instead of storing
a theoretically changing pointer to it inside each of the FWL widget's
properties, retrieve the one and only instance as needed. Then remove
code to set it into widgets, which mainly involved propogating the
values to sub-widgets to keep them in sync.

Change-Id: Ia645a338bdc600cb9199d439c09aabdbe959bfe8