Introduce class EmbedderTestEnvironment
Then initialize the library once per binary execution rather than
once per test. Should a clean library initialization be required for
a specific test, then unload/reload the library in that specific test.

Consequently, isolate creation takes place in the Environment, rather
than in each Test.

This lays the groundwork for making some of the XFA tests faster,
but more importantly will allow for a split in CFGAS between a
per-module singleton and a GC'd font mgr down the road.

Change-Id: I528005fbecd061453060ff33f077414020f5987f