Add CPE for libpng.
CPEs are the way that MITRE tracks different projects and versions, and
that's the input required by Vomit, Google's automated vulnerability
notification system. At present Vomit is unable to identify the versions
for these components and is therefore reporting the wrong CVEs in crbugs.

Adding this CPEPrefix line will enable Vomit to notify for any
vulnerabilities affecting these versions.

This is copied from https://crrev.com/746237

Change-Id: I0f04446763510ca4cc85746adb410ae9f3f4e135