Add FPDF_GetTrailerEnds() API
This follows the same pattern as FPDFSignatureObj_GetTime(), so the
client has to call this function twice, but allocation of the buffer
happens outside pdfium.

The benefit of this API is that client code can learn about the initial
trailer and the trailer at the end of incremental updates for signature
verification purposes.

This information can be used to write client code that emits no warnings
when a document is signed multiple times, but still warn when there are
unsigned incremental updates between the signatures or after the last
one. This behavior matches what Acrobat does, but it does not seem to be
in the PDF reference. So it is reasonable to have an API that allows
doing this, but not hardcode this policy into pdfium.

Change-Id: Ic741b672821d7743df88cdf51af54e85251857e6