Fix CPDFSecurityHandlerEmbedderTest.PasswordAfterGenerateSave for Skia.
When rendering text with AGG,  CFX_RenderDevice::DrawNormalText() round
the value of each glyph's horizontal position into an integer depending
on the |anti_alias| mode. However Skia renders with the float value of
each glyph's horizontal position. This results in the discrepancies on
text rendering for embeddertest
CPDFSecurityHandlerEmbedderTest.PasswordAfterGenerateSave.

Since Skia's rendering result is acceptable, this CL adds the expected
checksum value for Skia so that this test can pass.

Bug: pdfium:1500
Change-Id: I152f3cb672dcad16b593a55e2de67753b1167ba0