Use CheckedNumerics to calculate pitch in CalculatePitchAndSize().
In CFX_DIBitmap::CalculatePitchAndSize(), instead of manually avoiding
integer overflow when calculating the pitch, use CheckedNumerics instead
by instantiating a FX_SAFE_UINT32.

This raises the limit for the maximum possible pitch value in some
cases, such that the manual integer overflow check for the size value is
the limiting factor for successful calculations.

Bug: chromium:1124660
Change-Id: I83ec7fe8eff69cc4101dfecf4d8360bd9dea39f2