Rewrite some loops in CFXA_FMParser.
Resolve some TODO() about whether the loops are necessary. From looking
at the code, we've got a for-loop with a break as its last statement,
which looks unnecessary until one spots the continue statements lurking
within. So remove the continues and the final break to make this clearer.

OTOH, the TODO() might have been referring to whether the sequence of
tokens required to loop the loop could actually occur, which does seem
possible.

Change-Id: I64778892ae175b7db6a6455c92e666c48ec82ab5