Rework pdf_fuzzer_init_public.cc
Allow fuzzers to access globally-created resources, specifically
the cppgc garbage collector heap, for future fuzzing of GC'd
objects. This is complicated by the component build, where the
fuzzers's LLVMFuzzerTestOneInput() entry points live in the pdfium.so
library under different names, with the public init occurring in a
small shim above that.

-- Introduce a .h file for use by the fuzzers.
-- Rename class to match file (actual name unimportant)
-- Create an external heap when XFA is present.
-- Remove wrong comment about static variables being the same.

Change-Id: I97ba63dd19f5317f1635cbf3711c3551c3794852