Make char widths signed again.
Undo https://pdfium-review.googlesource.com/24410, because widths can be
negative for font glyphs. To fix https://crbug.com/806612, the correct
solution would have been to just delete the failing assertion. The CL
being reverted did that already, but for different reasons. (As
discussed on the code review, the assertion in question could not fail
when widths are unsigned.)

Bug: chromium:806612,chromium:1128284
Change-Id: Idca065ef9ade650613f8f39393d8513cedf31b6e