Fix FPDFText_GetCharBox() returning excessively tall char boxes.
In CPDF_TextPage::ProcessTextObject(), when the initial char box
calculation produces too small of a char box, the method compensates by
usng the font size for the height, and the glyph width for the width.
The height compensation code uses the font size in the wrong unit, so it
is 1000 times bigger than it needs to be. Fix this and update the
affected test case to show FPDFText_GetCharBox() no longer returns
excessively tall char boxes.

Bug: pdfium:1591
Change-Id: I1641b882c11c2f06b8a610c9d1ae11c850cd5edf