Use members in CFXA_TabParam sorting code
This is not essential for the first version of cppgc, since we are
not GCing while the stack is still active.

Also, make this a move-only type and avoid indirection through
unqiue_ptrs. The result is a move of three pointers instead of
one, but memory management becomes the container's problem, and
would be correct if we someday had a heap vector.

Change-Id: If76d1d0e782366d5a7d75e4aa4d89df3afe2e2a1