Garbage colllect CXFA_LocaleMgr.
Although not strictly required, this replaces a raw pointer to
an node with a traceable member.

In turn, pass FontMgrIface to the string formatter only on calls
that are affected by it. This prevents a lower layer that knows
nothing about garbage collection holding an unowned pointer to
a GC'd object.

Bug: pdfium:1563
Change-Id: Id6a372136f171a307b764a95d478c108cd3be5e4