Avoid LSAN leak reports in XFA fuzzers
Try #2. It looks like LLVM fuzzer may invoked leak detection in-between
any two runs of a fuzzer test case. We may be able to go back to the
periodic GC if we are willing to append detect_leaks=0 to the fuzzer
scripts, but ensure that this fixes the problem first and that there
isn't some funny business still lurking.

Bug: chromium:1131837
Change-Id: Id0a42857f3689494d928139f2cbc8966dfc42c68