Add a pixel test for a link annotation that does not render.
Add a pixel test which contains a link type annotation, which has "U"
(Underline) as its border style. This is to ensure border style "U" is
supported for rendering annotations.

The test expectation shows the current rendering result, which is blank.
But the correct rendering result should be an blue underline at the bottom
of the specified rectangular border box.

Bug: chromium:237527
Change-Id: Id1e93493d0d545b26d6d4dd1b3f6da67c20144ac