Move FPDFText_GetLooseCharBox() logic into cpdf_textpage.cpp.
Move the logic to GetLooseBounds() within cpdf_textpage.cpp and expose
it via CPDF_TextPage::GetCharLooseBounds(). This consistently structures
the code so the core logic is in core/, and the fpdfsdk/ code is there
to implement the public API. Also, GetLooseBounds() may get another
caller within cpdf_textpage.cpp in the near future.

To make this work, move part of CPDF_TextPage::GetCharFontSize() into a
standalone GetFontSize() function, for use with GetLooseBounds().

Change-Id: I72df08c730673a017ab18cc49108cf3423077fd8