Stop statically enforcing Xcode version.
The OSX machines in "luci.flex.try" pool that run pdfium tryjobs
have been upgraded from OSX 10.14 to OSX 10.15. Per [1],
the preferred Xcode version on OSX 10.15 is '12a7209' and most
tryjobs (from other projects) install that version. But when
a pdfium tryjob runs, it rolls Xcode back to '11b52'. Reinstalling Xcode
back and forth like that takes time (~20 min), it caused some tryjobs
in the "infra" project to timeout.

Unhardcode Xcode version in pdfium configs. It should allow the recipe
to pick the "best" version ('12a7209' in this case).

[1]: https://source.chromium.org/chromium/chromium/tools/depot_tools/+/master:recipes/recipe_modules/osx_sdk/api.py;l=21?q=11b52&ss=chromium

R=thestig@chromium.org, tandrii@google.com

Change-Id: I6f71f1efd8a4bdba720d308dbda03bd765008fc6