Avoid relocations in struct BarCodeInfo.
Per Chromium's docs/native_relocations.md, sometimes it makes sense to
restructure read-only data to avoid relocations. In this case, the
results are as follows on Linux:

.rodata: +1376
.data.rel.ro: -1488
.rela.dyn: -1488

So even disregarding the relocation data, which varies on platform, the
amount of read-only data goes down as a result of this change.

Change-Id: I3bf1312d5ec7ef5ac5da346b2e37bf2a22f15bd9