Improve font family name matching method.
When looking for substitute fonts for a given font family name,
CFX_FolderFontInfo::FindFont() simply checks whether this family name
is a substring of any installed font to decide whether an installed
font is a potential candidate. This will cause false matches when the
given family name is a substring of the installed font's family name.
For example, "Univers" is a substring of "Universal" but they are not
in the same font family.

This CL checks whether the next following character after the matching
substring is lower case to decide whether the installed font's family
name is fully covered. If the next character is lower case, then
consider it a family name mismatch.

Bug: chromium:1135949
Change-Id: If63c9eeb459671c084c0502154219738754f1ff8