Improve font family name matching method.
When looking for substitute fonts for a given font family name,
CFX_FolderFontInfo::FindFont() simply checks whether this family name
is a substring of any installed font to decide whether an installed
font is a potential candidate. This will cause the false match when the
given family name is a substring of the installed font's family name,
for example, "Univers" and "Universal" not in the same font family.

This CL checks the next following character after the matching substring
is lower case to decide whether we have fully covered the installed
font's family name. If the next character is lower case, then consider
it a family name mismatch.

Bug: chromium:1135949
Change-Id: If63c9eeb459671c084c0502154219738754f1ff8