Make FXDIB_format an enum class.
For didactic puprposes. Try to remove all arithmetic related to
computing members of this type, such that the only way to
get a particular value is via using its name.  Then we can
determine which names are never used.

This, of course, presumes that we don't actually have
types of dibs that don't have corresponding enums.  The enum
values are actually a combination of deptj + flags, and not
all permuations are named.  Sometimes we operate by name,
other times by properties. But so long as we don't cobble
these together by arithmetic, we should have closure over
the set of names.

There are two places where there may still be happening. We'd
need to prove that the flags are 0 in the spots so commented.

Nonetheless, i.e. both of these show no hits,
git grep '(FXDIB_Format)'
git grep 'cast<FXDIB_Format'

Change-Id: I3c5614bb7f71f64a98fc6a772b56bbab0a2dad91