Change Arial charset to CHARSET_FLAG_ANSI.
Arial is an ANSI font, therefore this CL changes its charset value to
CHARSET_FLAG_ANSI for testing. Due to this change, Arial is no longer a
mock-up symbol font in CFX_FolderFontInfoTest, and it should no longer
be a matching font for |kSymbol| (with charset of CHARSET_FLAG_SYMBOL).
Therefore, this CL updates the expectation for |kSymbol|'s matching
font as well.

Change-Id: Ib8ce8b193dca884926a07024baa7c746f4b7fe48