Change Arial font's charset value to CHARSET_FLAG_ANSI for testing.
This CL changes Arial font's charset value to CHARSET_FLAG_ANSI in
CFX_FolderFontInfoTest since Arial is in fact an ANSI font.

Due to this change, Arial is no longer a mock-up symbol font for the
test, and it should no longer be a match for |kSymbol|, of which the
charset value is CHARSET_FLAG_SYMBOL. Therefore, this CL also updates
the expectation for |kSymbol|'s matching font.

Change-Id: Ib8ce8b193dca884926a07024baa7c746f4b7fe48