Make FXDIB_format an enum class.
Try to remove all arithmetic related to computing enums of this
type, such that the only way to get a particular value is via
using its name.  Then we can determine which names are never used.

This, of course, presumes that we don't actually have
types of dibs that don't have corresponding enums.  The enum
values are actually a combination of depth + flags, and not
all permutations are named.  Sometimes we operate by name,
other times by properties. But so long as we don't cobble
these into existence by arithmetic, we should have closure over
the set of names.

There are two places where there may still be happening. We'd
need to prove that the flags are 0 in the spots so commented.

Nonetheless, i.e. both of these show no hits,
git grep '(FXDIB_Format)'
git grep 'cast<FXDIB_Format'

so there shouldn't be any funny business going on.

Even better, a grep for 1bppCmyk (which matches both the old and new
enum name) shows this is only brought into existence by the fuzzer.

Same for 8bppCmyk.  And 8bppCmyka can only be brought into existence
if 8bppCmyk exists, which we just said doesn't exist.

We can remove those in a follow-up.

Change-Id: I3c5614bb7f71f64a98fc6a772b56bbab0a2dad91