Make FXDIB_format an enum class.
Try to remove all arithmetic related to computing enums of this
type, such that the only way to get a particular value is via
using its name.  Then we can determine which names are never used.

This, of course, presumes that we don't actually have
types of dibs that don't have corresponding enums.  The enum
values are actually a combination of depth + flags, and not
all permutations are named.  Sometimes we operate by name,
other times by properties. But so long as we don't cobble
these into existence by arithmetic, we should have closure over
the set of names.

i.e. this now shows no hits,
git grep 'cast<FXDIB_Format'

so there shouldn't be any funny business going on.

Change-Id: I3c5614bb7f71f64a98fc6a772b56bbab0a2dad91