Disallow more CFX_ScanlineCompositor conversions.
Disallow the combination where the destination bitmap format uses
palettes and the reverse byte order bit is set. CFX_ScanlineCompositor
is not designed to handle this case, and will end up doing the wrong
thing. e.g. for a source format of `FXDIB_Argb` and destination format
of `FXDIB_8bppRgba`, it will call
CompositeRow_Rgb2Argb_NoBlend_NoClip_RgbByteOrder(), which is
completely nonsensical.

Bug: chromium:1139825
Change-Id: I6a4d7a402d4006d7454188cf9f66daa04ed98062