Remove `FXDIB_Format::kCmyk`.
No code ever creates bitmaps of this format, so any code that handle
this format is dead code. As a result of this removal,
`FXDIB_Format::kCmyka` is also unused and removed for the same reason.
Many utility functions that check to see if a bitmap is CMYK will be
removed in a follow-up CL.

Change-Id: Idab3556bf80ac1d49004c4de6cbb81aa5bb4d23b