Fix issue that some characters from the embedded font don't render.
Due to character composition, different unicodes could represent the
same character, which means they are mapped to the same CID. However the
current implementation for CID to unicode mapping is always 1:1, which
leads to the situation that certain characters cannot be recognized by
one of its valid unicodes during rendering.

Therefore, this CL changes the|m_Map| in CPDF_ToUnicodeMap to
std::multimap so that it can have multiple entries with the same CID as
keys.

This CL also changes |to_unicode| in LoadCompositeFont() to
std::multimap so that when embedding a font to a PDF file, all valid
unicodes for the same CID won't overwrite each other.

Change-Id: Ifadc2aa4df0de14e9d5a7c38da209f81769c0b3b