Move the Noto font file for testing to top-level third-party directory.
This CL moves the subset font file for NotoSansSC-Regular to PDFium's
top-level third-party directory, and changes the README.pdfium file to
comply with Chromium's guideline adding_to_third_party.md.

This CL also renamed the directory name to "NotoSansCJK/" so that more
Noto CJK fonts can be added in the same directory for increasing test
coverage in PDFIum embedder tests.

Bug: pdfium:1608
Change-Id: I28d59dceaa63e66d9a537d9d1485b307b2223167