Enable FPDFAnnotEmbedderTest.BUG_1206 for Skia/SkiaPaths.
With Skia/SkiaPaths enabled, the rendering result for
FPDFAnnotEmbedderTest.BUG_1206 has small discrepancies compared to AGG's
expected result. These discrepancies are caused by Skia's own way of
hairline handling which happens when the scale of rectangles are
relatively small.

Since the discrepancies are not very visible and the Skia/SkiaPaths
rendering results are considered acceptable, this CL updates the
Skia/SkiaPaths expectation for this test so that it can pass.

Bug: pdfium:1500
Change-Id: I8692d0469186d1a8b779f13810cdba6c53221805