Split up setters and getters
Because we hate out parameters.  We found one place where the logic was
inverted between setting and getting, which now becomes impossible when
getters are required to return results by value.

There's a small amount of duplicated code in some of the prologues
of these routines, but it is small and worth the trade-off for clairty.

Change-Id: I81fb68a34567957dca521c10440ae49c70b92886