FPDF_GetTrailerEnds: make this not depend on whitespace
'endstream' is typically followed by a newline, then we get 'endobj',
again followed by a newline. When CPDF_SyntaxParser::ReadStream() reads
the stream, then it reads 'endobj', and then resets the position back to
the end of 'endstream'. This mechanism is disabled in case there is
whitespace between the tokens and the newline, see the end of the
function.

This results in reporting no trailer ends, as the parsing fails, as the
next token is expected to be 'endobj', but it's the ID of the next
object instead.

Fix the problem by relaxing the conditions in FPDF_GetTrailerEnds(), so
we handle when 'endobj' is already consumed.

The alternative would be to make the position at the end of

CPDF_SyntaxParser::ReadStream() more deterministic (not depend on
whitespace), but that may cause unwanted side-effects.
Change-Id: I7048e8d081af04af3dd08d957212c885b7982b5e